Cleanup Files
Moved several files out of asset folder and pushed into appropriate folders.
