+++ Bean Man Gets His Cool Back Unity Proj/Assets/Prefabs.meta
@@ -1,5 +1,6 @@
 fileFormatVersion: 2
-guid: fb01be13d6e88ca488dda82150319bfc
+guid: 4b4c2eaab30a3944cbee1ecf97da89ca
+folderAsset: yes
 DefaultImporter:
   externalObjects: {}
   userData:
