+++ Bean Man Gets His Cool Back Unity Proj/Assets/Scripts/AnimationSounds.cs.meta
