+++ Bean Man Gets His Cool Back Unity Proj/Assets/Sprites/Static Screens.meta
@@ -1,5 +1,6 @@
 fileFormatVersion: 2
-guid: 560a88da2bbc70140bed167f0ba7fe37
+guid: 8e3e47cdb0102ba4ca0d1d0f61d2a114
+folderAsset: yes
 DefaultImporter:
   externalObjects: {}
   userData:
