+++ Bean Man Gets His Cool Back Unity Proj/Assets/Sprites/Static Screens/Beango_Screen1.png.meta
