+++ Bean Man Gets His Cool Back Unity Proj/Assets/Sprites/Static Screens/Beango_Screen2.png.meta
